testing added

+++ src/main/java/system/university/pds/model/Classroom.java
@@ -2,9 +2,7 @@ package system.university.pds.model;
 
 import lombok.Getter;
 
-import java.math.BigInteger;
-
-public record Classroom(@Getter BigInteger classroomNumber, short capacity) {
+public record Classroom(@Getter int classroomNumber, short capacity) {
     public Classroom {
         if (capacity <= 0) {
             throw new IllegalArgumentException("Capacity must be greater than 0");
